+++ phpmyfaq/translations/language_en.php
@@ -1486,5 +1486,6 @@ $PMF_LANG['msgPluginDescription'] = 'Description';
 
 // added v4.1.0-alpha.2 - 2025-02-23 by Thorsten
 $PMF_LANG['msgReportABug'] = 'Report a bug';
+$PMF_LANG['msgDashboard'] = 'Dashboard';
 
 return $PMF_LANG;