+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This is a log of major user-visible changes in each phpMyFAQ release.
 - added plugin administration backend (Thorsten)
 - added experimental support for PDO (Thorsten)
 - added Urdu translation (Thorsten)
+- updated Spanish translation (Thorsten)
 - improved online update feature (Thorsten)
 - migrated from WYSIWYG editor from TinyMCE to Jodit Editor (Thorsten)
 - migrated from JavaScript to TypeScript (Thorsten)